requirements.txt
torch
torchvision
opencv-python
pandas
numpy
tqdm
datasets
Pillow
